PT-6833 OlapLeaderboardPeriod -> OlapPeriod in new API, move new periods there
We cannot easily add new periods to OlapLeaderboardPeriod, because it is used
in v0.1 API. They will try to access tables partitioned by period,
which will break. So we need a new enum for period.

But there's already OlapPeriod. It is currently unused:
the implementation of the only controller that is spec'd to use it
actually ignores this parameter. So we are adding the new period there,
and adding the periods from OlapLeaderboardPeriod that were not there.

Changing existing v2.0 API to OlapPeriod is backwards compatible
for all periods except for the new ones.
But they are unused yet, so it shouldn't be a big problem.